dues update

+++ app/src/main/java/com/marufalam/dufa/data/models/payRenew/RequestPayRenew.kt
@@ -3,7 +3,7 @@ import com.google.gson.annotations.SerializedName
 
 data class RequestPayRenew(
     @SerializedName("amount")
-    val amount: Int?,
+    val amount: Double?,
     @SerializedName("membership")
     val membership: String?,
     @SerializedName("userinfo_id")
